pyproject.toml
[tool.poetry]
name = "stardewaccounting"
version = "0.1.0"
description = ""
authors = ["Eric Batista <klose.eric31@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.87.0"
uvicorn = "^0.19.0"
pydantic = "^1.10.2"
python-dotenv = "^0.21.0"
pytest = "^7.2.0"
streamlit = "^1.14.1"
alembic = "^1.8.1"
psycopg2-binary = "^2.9.5"


[tool.poetry.group.dev.dependencies]
isort = "^5.10.1"
black = "^22.10.0"
autoflake = "^1.7.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
